Bug fix: Suppress "false" error messages from jive.ui.Textarea & jive.ui.Textinput
The log is often 'spammed' with 'get_image_surface - no SDL surface available'
error messages when an empty text line is drawn into a Textarea, or when
drawing 0 length text strings into a Textinput area.

These messages are intrusive, and are really false alarms. They can make it
more difficult to identify which log entries actually matter.

This change ensures that no attempt is made to draw these 0 length text
strings, because there is nothing to draw. This eliminates the error messages.