Bug fix: Check network diagnostic hangs if LMS not completely discovered
If the LMS server port (generally 9000) has not been set, then the
"Port 9000 test" will lock up the UI in a "Check network" spinny. This change
prevents that by checking that the server port has been set to a numeric
value, and failing the check if it has not.

This is unlikely to be encountered often, but it can happen. In this
circumstance the LMS server port (normally, but not always, 9000) may still
be set at its initial value of 'false', or perhaps undefined. The call to
'SocketTcp' will fail by way of an '_assert', and the UI will lock up in a
spinny.

Observation: This can happen if a Radio's local player has been initialized
by virtue of its holding a routable ip address for an LMS which lives on a
different network. Automatic server discovery can only succeed on its own
network, so that process will not have populated the server port.

Most users in this scenario will have already set up a manual LMS address
through the Choose Music Source applet CMS, and would not encounter the issue.