Update SocketHttp.lua
This call t_setResponseBody with nil is treated as reporting an error.
But the code is executed when the request is completed.

I don't think there is any need to call. t_setResponseBody at all.

Current code prevents PatchInstaller.lua from working when the patch URL is redirected.  When the 301 header is read, this code causes RequestHttp to attempt to report an error.  Since there is no "code" in nil, line 406 in RequestHttp.lua attempts to concatenate a string with a boolean (false) - an error.

RequestHttp then closes the sink, which should be left for use by the redirect.

I think this is the correct place to fix the problem.  A workaround can be added in RequestHttp.lua to handle this case, but there is no reason to call t_setResponseBody with nil in the first place.