Retrieve a default TimeZone from a connected LMS server
This change provides SqueezeOS with the ability to obtain a default TimeZone
from a connected LMS server if its TimeZone has not been initialized. This
would be the case on first set up, or following a factory reset.

This service was formerly provided by MySqueezeBox.com, but that is no longer
available.

The relevant functionality is now to be provided by the LMS server, which will
host a new '/tz' HTTP endpoint. SqueezeOS issues a simple HTTP 'GET' to this
endpoint.

The server itself will obtain the TimeZone from
'https://stats.lms-community.org/api/time'.

The server is expected to respond by returning an Olson formatted TimeZone
string, with a '200' (RC_OK) response code. If the server fails to obtain a
valid TimeZone it will simply return a '404' (RC_NOT_FOUND) response.

Older versions of the server, which will not be equipped with the 'tz'
endpoint, will also return '404'.