setTimezone: Check data file exists before setting tz
This change adds some checks to guard against processing an
inappropriate Olson TimeZone string.

The checks ensure that an underlying tz data file appears to be present
in the system, and that the tz data file falls within the
'/usr/share/zoneinfo' hierarchy.

The checks are carried out on the canonicalized form of the target tz
data file. This assures that any symbolic link within the zoneinfo
hierarchy can be resolved to its underlying tz data file, and protects
against unexpected path traversals.